Fix a panic when comparing types where the underlying type is string
The code was checking if the type's _kind_ was string and then attempting to
do a type assertion of `var.(string)`. However, this assertion only works if
the type actually _is_ a string. If it's some other type (like `type foo
string`) then this ends up giving us a runtime panic.