Update the LICENSE to match the 'standard' MIT license file
The reason is that since the name of the license is not stated
explicitly and the license text differed very slightly I had to perform
a diff to ensure it actually is MIT licensed. Also GitHub does not
detect the license for this project currently (probably for the same
reason).

Changes:

* Remove the line 'Please consider promoting this project...' that is
  not part of the MIT license
* Use the same formatting as choosealicense.com
* Explicitly state 'MIT License'
* Update the copyright year