Add Drop method to Mock
Mock.Drop removes a Call from the slice of ExpectedCalls on the Mock.

This is helpful when repeatedly mocking the same function with the same
arguments at different points in the test but with different returned
results as at the moment doing this adds new Calls to the ExpectedCalls
slice even if the return arguments are different.