Preserve stack frame for mock parent method call
Go tip contains following commmit, that inlines function with single
call bodies.

https://github.com/golang/go/commit/13baf4b2cd34dfb41c570e35b48ec287713f4d7f

`(*Call).On()` is the exact target for the improvement in Go repo. Due to
the inlining, assert.CallerInfo() can't not detect the file and line
number of the call to `(*Mock).On()` from `(*Call).On()`. Thus, the test
fails.

Adding the compiler directive `go:noinline` prevent this effect and make
mock package works with go tip as before.