Prepare to switch call stacks when retrieving stack traces
For each test failure, CallerInfo() computes a stack trace to include
in the output. For each stack frame we call runtime.Caller().

This approach misses useful information when using Go subtests.
This is because a call like

	t.Run("My subtest", func(t *testing.T) {
		assert.Fail(t, "something went wrong")
	})

merely enqueues the given function but does not run it straight
away. Only after all subtests have been added, the function is actually
called, and by then, the t.Run (and anything above) is no longer part
of the stack trace[*].

To add this information to the stack traces, we need to special-case
subtest invocations as we climb up the call stack. Whenever we
encounter a subtest invocation, we want to continue with the call
stack of the call to t.Run() that enqueued the subtest.

This means we can't use runtime.Caller() anymore to walk the
stack, because that only walks the current call stack.  Switch to
runtime.CallersFrames() (introduced in Go 1.17, 2016), to allow the
next patch to implement the fix described in the previous paragraph.

[*] When a function has many subtests, it is very helpful if my editor
can jump to the failing subtest.

Co-authored-by: Johannes Altmanninger <johannes.altmanninger@symflower.com>