assert: Fail `(Not)Subset` when map superset is missing elements (#1270)
When a subset map contains an item with a key that is not present in
the superset map, attempts to reflect the interface of the value at
the absent key cause panics, which return false on recovery but do not
properly fail the test with a message.

This also aligns the `Subset` failure message with the corresponding
`NotSubset` failure message.