Fix calling Unset on call with argumentMatcher.
When calling Unset on an call that contains an argumentMatcher as an expected value, the call fails with panic, fails the test and any remaining Expectations remain in place causing the test to fail with messaging about failed expectations. To work around this, during Diff check, compare the matcher to the incoming actualMatcher if it is an argumentMatcher instead of attempting to match it as if it was a received argument.