ci: improve github workflows
Centralise the CI in main.yml making it easier for contributors to
understand the requirements avoid them to view non-standard script
files.

Fix detection of new files so that broken changes are detected by CI.

Eliminate duplicate test run to reduce the time needed to run.

Add check for go mod tidy to ensure that go mod is clean before it's
allowed to be merged.

Update to action/setup-go@v4 which enables caching by default improving
the performance of all workflows.