assert: handle TokenTooLong error scenario while formatting assertion messages
As pointed out in #1525, when the assertion message is too long, it gets
completely truncated in the final output. This is because,
`bufio.Scanner.Scan()` has a default `MaxScanTokenSize` set to `65536`
characters (64 * 1024). The `Scan()` function return false whenever the
line being scanned exceeds that max limit. This leads to the final
assertion message being truncated.

This commit fixes that by manually setting the internal scan buffer size
to `len(message) + 1` to make sure that above scenario never occurs.

Fixes #1525