Assertion for minimum number of method/function calls
There are cases where exact number of calls that will be made inside a method cannot be predtermined accurately. Eg, number of network calls made by a retry logic which is governed by a time out. Number of calls made in such cases will depend on time of previous call and time out of governing context. Hence ability to ensure a minimum threshold is required