mock.AssertExpectationsForObjects fix panic with wrong testObject type.
Previously passing a mock.Mock by value rather than by reference worked (in cases without mutex locking issues) and logged a warning. This was broken by #1212 which introduced a breaking change in an attempt to fix go vet.

There is no clean way to fix the breaking change as we now have (and want) go vet in our CI.

This PR does not revert the breaking change but changes the panic to a test failure with a useful message.