Ensure deterministic order when diffing maps
When generating the diff for two maps where the map key is not a basic
type, the order is non-deterministic which makes the diffs unreadable.

By enabling the SpewKeys setting, the map keys will be spewed to strings
for the purpose of sorting them. This results in a deterministic order
and makes the diffs readable.