Add CollectT.Halt() as a fail-fast helper for EventuallyWithT
So fatal assertions invoked inside EventuallyWithT can stop the retry
loop.

Add the new HaltT struct that acts as a CollectT wrapper. Whenever an
error or failure is recorded, the wrapping HaltT will set CollectT's new
halt flag, so that the EventuallyWithT loop can detect this and error
out early.

This allows for cleanly breaking out of the test loop before the timeout
has elapsed, if necessary.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>